dcgan_generator/requirements.txt
chainer==1.19.0
cycler==0.10.0
Cython==0.25.2
filelock==2.0.7
matplotlib==2.0.2
nose==1.3.7
numpy==1.11.3
Pillow==3.4.2
protobuf==3.1.0.post1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
